dhcp: Cancel and cleanup tasks in DHCP server detector
When using the server detector as a Python module repeatedly, Python was
throwing errors that tasks were destroyed that were still pending. This
commit implements some cleanup of tasks so that this leaking of pending
tasks is prevented. I'm not sure if this catches everything, but it
solves it for my use case where I repeatedly call detect_servers().